finish interface refactor
