+++ frontend/interfaces/Data.tsx
@@ -1,0 +1,7 @@
+import { NewEvent } from "./Events"
+import { FilterTypes } from "./Filters"
+
+export interface Data {
+  eventData: Array<NewEvent>,
+  filters: FilterTypes
+}
